ctfs/NN0x04/uae/src/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1.1
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.14.1
redis==3.4.1
